+++ feature/main/src/main/java/kr/co/main/calendar/model/FarmWorkModel.kt
@@ -1,4 +1,4 @@
-package kr.co.main.model.calendar
+package kr.co.main.calendar.model
 
 import kr.co.domain.entity.type.FarmWorkType
 import kr.co.domain.entity.type.FarmWorkEraType
