+++ feature/main/src/main/java/kr/co/main/calendar/providers/FakeScheduleModelProvider.kt
@@ -1,15 +1,12 @@
 package kr.co.main.calendar.providers
 
-import android.os.Build
-import androidx.annotation.RequiresApi
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
-import kr.co.main.model.calendar.CropModel
-import kr.co.main.model.calendar.ScheduleModel
-import kr.co.main.model.calendar.type.CropModelColorType
-import kr.co.main.model.calendar.type.CropModelType
-import kr.co.main.model.calendar.type.ScheduleModelType
+import kr.co.main.calendar.model.CropModel
+import kr.co.main.calendar.model.ScheduleModel
+import kr.co.main.calendar.model.type.CropModelColorType
+import kr.co.main.calendar.model.type.CropModelType
+import kr.co.main.calendar.model.type.ScheduleModelType
 import java.time.LocalDate
-import java.time.LocalDateTime
 
 internal class FakeScheduleModelProvider : PreviewParameterProvider<ScheduleModel> {
     override val values = sequenceOf(
