+++ feature/main/src/main/java/kr/co/main/calendar/ui/calendarScreen/addDiaryScreen/diary_input/DiaryInputWrapper.kt
@@ -9,8 +9,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import kr.co.main.calendar.adddiary.diaryinput.DiaryWorkInput
-import kr.co.main.model.calendar.DiaryModel
-import kr.co.main.model.calendar.type.WorkDescriptionModelType
+import kr.co.main.calendar.model.DiaryModel
+import kr.co.main.calendar.model.type.WorkDescriptionModelType
 import kr.co.ui.theme.typo
 import java.time.LocalDate
