+++ feature/main/src/main/java/kr/co/main/calendar/ui/calendarScreen/calendarScreen/scheduleTab/ScheduleTabViewModel.kt
@@ -12,10 +12,10 @@ import kr.co.domain.usecase.calendar.GetHolidaysUseCase
 import kr.co.main.mapper.calendar.CropModelMapper
 import kr.co.main.mapper.calendar.FarmWorkModelMapper
 import kr.co.main.mapper.calendar.HolidayModelMapper
-import kr.co.main.model.calendar.CropModel
-import kr.co.main.model.calendar.FarmWorkModel
-import kr.co.main.model.calendar.HolidayModel
-import kr.co.main.model.calendar.ScheduleModel
+import kr.co.main.calendar.model.CropModel
+import kr.co.main.calendar.model.FarmWorkModel
+import kr.co.main.calendar.model.HolidayModel
+import kr.co.main.calendar.model.ScheduleModel
 import kr.co.ui.base.BaseViewModel
 import timber.log.Timber
 import java.time.LocalDate
