+++ feature/main/src/main/java/kr/co/main/calendar/ui/calendarScreen/searchDiaryScreen/SearchDiaryScreenViewModel.kt
@@ -4,9 +4,9 @@ import android.os.Parcelable
 import androidx.lifecycle.SavedStateHandle
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
-import kr.co.main.model.calendar.CropModel
-import kr.co.main.model.calendar.DiaryModel
-import kr.co.main.model.calendar.type.CropModelType
+import kr.co.main.calendar.model.CropModel
+import kr.co.main.calendar.model.DiaryModel
+import kr.co.main.calendar.model.type.CropModelType
 import kr.co.main.navigation.CalendarNavGraph
 import kr.co.ui.base.BaseViewModel
 import java.time.LocalDate
