+++ feature/main/src/main/java/kr/co/main/calendar/ui/common/content/CalendarBaseCard.kt
@@ -23,8 +23,8 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
-import kr.co.main.model.calendar.DiaryModel
-import kr.co.main.model.calendar.ScheduleModel
+import kr.co.main.calendar.model.DiaryModel
+import kr.co.main.calendar.model.ScheduleModel
 import kr.co.main.calendar.providers.FakeDiaryModelProvider
 import kr.co.main.calendar.providers.FakeScheduleModelProvider
 import kr.co.nbdream.core.ui.R
