+++ feature/main/src/main/java/kr/co/main/mapper/calendar/CropModelMapper.kt
@@ -3,9 +3,8 @@ package kr.co.main.mapper.calendar
 import kr.co.common.mapper.BaseMapper
 import kr.co.domain.entity.CropEntity
 import kr.co.domain.entity.type.CropType
-import kr.co.main.model.calendar.CropModel
-import kr.co.main.model.calendar.type.CropModelColorType
-import kr.co.main.model.calendar.type.CropModelType
+import kr.co.main.calendar.model.CropModel
+import kr.co.main.calendar.model.type.CropModelType
 
 internal object CropModelMapper
     : BaseMapper<CropEntity, CropModel>() {
