+++ feature/main/src/main/java/kr/co/main/mapper/calendar/CropModelTypeMapper.kt
@@ -2,7 +2,7 @@ package kr.co.main.mapper.calendar
 
 import kr.co.common.mapper.BaseMapper
 import kr.co.domain.entity.type.CropType
-import kr.co.main.model.calendar.type.CropModelType
+import kr.co.main.calendar.model.type.CropModelType
 
 internal object CropModelTypeMapper
     : BaseMapper<CropType, CropModelType>() {
