+++ feature/main/src/main/java/kr/co/main/mapper/calendar/HolidayModelMapper.kt
@@ -2,7 +2,7 @@ package kr.co.main.mapper.calendar
 
 import kr.co.common.mapper.BaseMapper
 import kr.co.domain.entity.HolidayEntity
-import kr.co.main.model.calendar.HolidayModel
+import kr.co.main.calendar.model.HolidayModel
 
 internal object HolidayModelMapper
     : BaseMapper<HolidayEntity, HolidayModel>() {
