+++ feature/main/src/main/java/kr/co/main/mapper/calendar/WorkDescriptionModelMapper.kt
@@ -3,8 +3,8 @@ package kr.co.main.mapper.calendar
 import kr.co.common.mapper.BaseMapper
 import kr.co.domain.entity.DiaryEntity
 import kr.co.domain.entity.type.WorkDescriptionType
-import kr.co.main.model.calendar.DiaryModel
-import kr.co.main.model.calendar.type.WorkDescriptionModelType
+import kr.co.main.calendar.model.DiaryModel
+import kr.co.main.calendar.model.type.WorkDescriptionModelType
 
 internal object WorkDescriptionModelMapper
     : BaseMapper<DiaryEntity.WorkDescriptionEntity, DiaryModel.WorkDescriptionModel>() {
